requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
detect-secrets>=1.5.0
pyyaml>=6.0
